+++ browser_use/browser/browser.py
@@ -82,9 +82,7 @@ class Browser:
 				'--disable-features=IsolateOrigins,site-per-process',
 			]
 
-	async def new_context(
-		self, config: BrowserContextConfig = BrowserContextConfig()
-	) -> BrowserContext:
+	async def new_context(self, config: BrowserContextConfig = BrowserContextConfig()) -> BrowserContext:
 		"""Create a browser context"""
 		return BrowserContext(config=config, browser=self)
 
@@ -151,7 +149,7 @@ class Browser:
 			stdout=subprocess.DEVNULL,
 			stderr=subprocess.DEVNULL,
 		)
-    
+
 		# Attempt to connect again after starting a new instance
 		for _ in range(10):
 			try:
